Fix multi-dimensional array index reaching (#2666)

+++ kernel/object.c
@@ -1045,8 +1045,9 @@ int zephir_update_property_array_multi(zval *object, const char *property, uint3
 			va_start(ap, types_count);
 			switch (types[0]) {
 				case 's': {
-					char *str = va_arg(ap, char*);
-					int len   = va_arg(ap, int);
+					char *str  = va_arg(ap, char*);
+					/* SL() pushes a size_t; see kernel/array.c. */
+					size_t len = va_arg(ap, size_t);
 					ZVAL_STRINGL(&offset, str, len);
 					break;
 				}
